tools/macro-scanner/requirements.txt
requests>=2.31.0
mysql-connector-python>=8.0.0
python-dotenv>=1.0.0
yfinance>=0.2.36
anthropic>=0.40.0
